pyproject.toml
[project]
name = "workspace-manager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pyobjc-framework-applicationservices>=12.1",
    "pyobjc-framework-cocoa>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "pyyaml>=6.0.3",
]
